Updates! (BACKUP)

+++ Source/Utilities.cpp
@@ -1091,10 +1091,10 @@ namespace DiscordCoreAPI {
 		return std::string("\033[1;40;91m");
 	}
 
-	bool nanoSleep(int64_t ns) {
+	bool nanoSleep(uint64_t ns) {
 #ifdef _WIN32
 		HANDLE timer = CreateWaitableTimerExW(NULL, NULL, CREATE_WAITABLE_TIMER_HIGH_RESOLUTION, TIMER_ALL_ACCESS);
-		LARGE_INTEGER largeInt{ .QuadPart = -ns / 100 };
+		LARGE_INTEGER largeInt{ .QuadPart = -static_cast<int64_t>(ns) / 100 };
 		if (!timer) {
 			return false;
 		}
